requirements.txt
# Web framework
streamlit==1.31.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# Azure services
azure-cosmos==4.5.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.2.0
opentelemetry-instrumentation-streamlit==0.1.0

# Azure OpenAI
openai==1.12.0

# API requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# Logging and monitoring
structlog==24.1.0